% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_central_tables.R
\name{get_central_tables}
\alias{get_central_tables}
\title{Get Names of Central Tables}
\usage{
get_central_tables()
}
\value{
Character vector
}
\description{
Get names of the non-linguistic variant
tables to be able to remove erroneous tables
when desired.
}
