% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_setup.R
\name{run_setup}
\alias{run_setup}
\title{Load LOINC Complete Package}
\usage{
run_setup(
  conn,
  conn_fun = "pg13::local_connect()",
  schema = "loinc",
  path_to_csvs,
  include_linguistic_variants = FALSE,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  log_schema = "public",
  log_table_name = "setup_loinc_log",
  log_csv_version,
  log_release_date
)
}
\arguments{
\item{conn}{Connection to a Postgres database.}

\item{conn_fun}{Connection expression.}

\item{schema}{Target schema for the LOINC load, Default: 'loinc'}

\item{path_to_csvs}{Path to the unpacked LOINC download.}

\item{include_linguistic_variants}{Include \code{LinguisticVariant} tables?}

\item{log_schema}{Schema for the table that logs the process, Default: 'public'}

\item{log_table_name}{Name of log table, Default: 'setup_loinc_log'}

\item{log_csv_version}{(Required) Version number povided in the download page and source zip file name.}

\item{log_release_date}{(Required) Release date provided in the download page.}
}
\description{
The CSV files unpacked from "LOINC and Relma Complete Download"
(\url{https://loinc.org/downloads/}) zip are loaded into Postgres.The
file version and release dates are taken along with the table names and row
counts are logged to a separate table after the loading is completed.
}
